Bug fix for false positives with no-multi-comp
Previously, when createElement was destructured from React, it would
cause any function defined in the file to be flagged as a React
Component. This change makes it such that the function must call
createElement to be considered a component.

Fixes #1088

---

Review: Use object.assign with sourceType: module in the added test over
babel-eslint.