Check parentheses with new lines for multiline JSX
This changes the options from a boolean to an enum with five possible
values: true, false, `ignore`, `parens`, and `parens-new-line`. True and
false are there for backwards compatibility. Using `ignore` equates to
using `false` while `parens` is equal to `true`. The new option
(`parens-new-line`) will check that the parentheses around the multiline
JSX are on separate lines for better readability.